[wmco] Add WMCO version to Windows node
This commit adds the version of the WMCO instance that configured each
node to that node's annotations. This will be used in the WMCO upgrade
process, as this allows WMCO to see which nodes were not configured by
the current version.

In addition a node controller was added to prevent the version
annotation being from changed to an unexpected value.