[nodeconfig] Fix network config script
The network configuration script contained bag logic to determine if the
cni config had the correct contents. The 'expected' config does not
contain carriage returns, while the actual created config file does.
Because the script did not account for this, the script would always
determine that the cni configuration needed to be recreated with the
correct contents. This can cause issues with containerd, which needs to
reload the configuration everytime this happens.